\name{D7905}
\alias{D7905}
\docType{data}
\title{Historical Data}
\description{
Historical drop test data where testers did not immediately terminate testing.
Useful for comparing modern test methods and simulation output.
}
\usage{data("D7905")}
\format{
  A data frame with 10 observations on the following 2 variables.
  \describe{
    \item{\code{P}}{a numeric vector}
    \item{\code{SD}}{a numeric vector}
  }
}
\source{
NASA Technical Note NASA-TN D-7905. (1970)
Written by J. B. Gayle.
\url{https://ntrs.nasa.gov/archive/nasa/casi.ntrs.nasa.gov/19750014413.pdf}
}
\examples{
data(D7905)
}
\keyword{datasets}