% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/droptest.R
\docType{package}
\name{droptest}
\alias{droptest}
\alias{droptest-package}
\title{droptest: Simulates LOX/GOX drop testing.}
\description{
Provides seven functions. Generates simulated data representing the LOX/GOX
drop testing process (Impact testing). Functions are provided to simulate 
trials, test series, and groups of test series. Functions for creating plots
specific to this testing method are also included. Test attributes and 
criteria can be set arbitrarily.
}
\details{
Drop testing -- sometimes called impact testing -- is used to evaluate if a
material will interact with liquid (LOX) or gaseous oxygen (GOX). The
material is exposed to the LOX/GOX and an impactor is dropped onto the
sample. Each drop is a bernoulli trial where a reaction is a failure and a
non-reaction is a success. The specified number of trials -- until failure --
completes one test.

While fundamentally a binomial process, drop testing -- performed by the
military and NASA -- yields results that are difficult to analyze. Numerous
tech briefs and standards have attempted to address the problem (see below
for more information). Testing stops as soon as a failure condition is
reached. If the failure condition occurs on drop one or two -- depending on
the failure criteria -- then the test returns only one or two result values.
Alternatively, if a material passes, or if the failure condition occurs on
the last observation, then the test returns as many result values as 
observations.

Simulation can be used to examine the behavior of this test procedure.

Inspired by NASA Technical Note "Computer Simulation of Threshold Sensitivity
Determinations" NASA-TN-D-7663. 1974.
\url{https://ntrs.nasa.gov/archive/nasa/casi.ntrs.nasa.gov/19750004618.pdf}

\strong{Note:} This work is not endorsed by or affiliated with NASA. Released
under MIT license.
}
\references{
Pass/Fail criteria and number of observations required have been defined in
 the following standards:
\itemize{
 \item NASA-STD-6001B
 \url{https://spaceflightsystems.grc.nasa.gov/SpaceDOC_II/Standards/documents/NASA-STD-6001B-1.pdf}

 \item ASTM D2512 \url{https://www.astm.org/Standards/D2512.htm}

 \item ASTM G86-17 \url{https://www.astm.org/Standards/G86.htm}
}

\strong{Note:} This package is not constrained by any standard. Arbitrary
test criteria and observations can be specified for maximum flexibility.

For more information on drop testing: 

\itemize{
 \item "An Instrument for Determination of Impact Sensitivity of Materials in 
 Contact with Liquid Oxygen" (AB6002-EB). Lucas (1960). 
 \url{https://www.astm.org/DIGITAL_LIBRARY/STP/MMR/PAGES/AB6002-EB.htm}

 \item "Reproducibility of Liquid Oxygen Impact Test Results"
 (NASA-TN-D-7905). Gayle (1970).
 \url{https://ntrs.nasa.gov/archive/nasa/casi.ntrs.nasa.gov/19750014413.pdf}

 \item "Lox/Gox Mechanical Impact Tester Assessment" (TM-74106). Bransford,
 Bryan, Frye, Stohler (1980).
 \url{https://ntrs.nasa.gov/archive/nasa/casi.ntrs.nasa.gov/19800006920.pdf}
}
}
\seealso{
\code{\link{dtrials}}
 \code{\link{dseries}}
 \code{\link{dgroups}}
 \code{\link{trialdev}}
 \code{\link{gayleplot}}
 \code{\link{ratioplot}}
}
\author{
Chad Ross \email{chad.ross@gmail.com}
}
