% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtrials.R
\name{dtrials}
\alias{dtrials}
\title{Completes a simulated drop test.}
\usage{
dtrials(q, max.trials = 20, fail.criteria = 1,
  data.structure = "data.table")
}
\arguments{
\item{q}{Integer. Specifies the probability that a reaction occurs. 
A reaction is interpreted as a failure therfore q is the probability
of failure for the bernoulli trials. Probability of success would be
\eqn{p = 1 - q}. See \url{https://en.wikipedia.org/wiki/Bernoulli_trial}}

\item{max.trials}{Integer. The maximum number of bernoulli trials to
perform. where each trial represents one drop of the impactor onto
a sample. The simulated test can only reach this number of trials if
no reactions (failures) occur. In other words: trials performed will
always be less than or equal to \code{max.trials}. Default is 20.}

\item{fail.criteria}{Integer. Specifies number of reactions (failures)
that can occur before an entire test is considered a failure. Default
is 1.}

\item{data.structure}{Instructs function to return result as a data.table
or a list. Default is data.table. 

\strong{Note:} Other functions in this package only work with
data.tables. List is an option strictly for future flexibility.}
}
\value{
Result of simulation as a data table or list -- depending on value
  of \code{data.structure}. 
  
  Where:
   \strong{F_CRITERIA} is the failure criteria specified (default is 1).
   \strong{REACT} is the total number of reactions (failures).
   \strong{NON_REACT} is the total number of non-reactions (successes)
   \strong{TRIALS} is the number of trials performed until failure 
    condition met. If the failure condition was met. If the failure
    condition was not met then this values will always be equal to
    MAX_TRIALS.
   \strong{MAX_TRIALS} is the maximum number of trials to perform as
    specified (default is 20). TRIALS will always be less than or equal to
    MAX_TRIALS.
   \strong{PCT_REACT} is the percent of trials that yielded a reaction
    (failure).
   \strong{Q} is the probability of failure (reaction) as specified.
   \strong{P} is the probability of success (non-reaction).
   \strong{RESULT} is whether the test series as a whole failed or passed
    based on the failure criteria specified.
}
\description{
\code{dtrials} returns bernoulli trials that compose a drop test.
}
\examples{
  dtrials(0.05)
  dtrials(0.05, max.trials = 60)
  dtrials(0.05, fail.criteria = 2)
  dtrials(0.05, max.trials = 60, fail.criteria = 2)
  dtrials(0.05, data.structure = "list")
}
