\name{PREGENTD}
\alias{PREGENTD}
\docType{data}
\title{Trial Deviation for Pre-Generated Data}
\description{
Output of \code{trialdev()} on a large pre-generated dataset.

Original dataset represents fifty levels of q where each level has 100,000
simulated test series. For a total of 5,000,000 simulated tests.
}
\usage{data("PREGENTD")}
\format{
  A data frame with 50 observations of 4 variables.
}
\seealso{
  \code{\link{dseries}}
  \code{\link{dgroups}}
}
\examples{
data(PREGENTD)
}
\keyword{datasets}