% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trialdev.R
\name{trialdev}
\alias{trialdev}
\title{Calculates the "Trial Deviation".}
\usage{
trialdev(sim.values)
}
\arguments{
\item{sim.values}{data.table. Data table produced by droptest::trials, 
droptest::series, or droptest::groups.}
}
\value{
Data table of q, p, trial deviation, and average total trials per
  test (average distance from q). Aggregated by q levels.
  
 Where:
 \strong{Q} is the probability of failure (reaction) as specified.
 \strong{P} is the probability of success (non-reaction).
 \strong{TRIAL_DEV} is the average distance from q for the total percent of
  reactions (failures).
 \strong{AVG_TRIALS} is the average number of simulated trials for each
  level of q.
}
\description{
\code{trialdev} Creates data.table with the average distances from q for
the total percent of reactions (failures).
}
\examples{
  trialdev(dtrials(q = 0.05, max.trials = 60, fail.criteria = 2))

}
\seealso{
\code{\link{dtrials}}
 \code{\link{dseries}}
 \code{\link{dgroups}}
}
