% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgroups.R
\name{dgroups}
\alias{dgroups}
\title{Generates groups of simulated test series.}
\usage{
dgroups(num.groups = NULL, probs = NULL, ...)
}
\arguments{
\item{num.groups}{Integer. Specifies how many groups of drop tests to
produce.}

\item{probs}{Vector. Specifies values of probabilities to be used when 
multi.p is TRUE. Note: Vector length must equal value of num.groups.}

\item{...}{Passes values to \code{dseries}.}
}
\value{
Data table of groups where each group consists of multiple drop
  tests. \strong{Note:} Only data table is supported at this time.
}
\description{
\code{dgroups} returns a collection of multiple simulated test series. Each
batch of test series are identified with a group number. Within each group
test parameters will be identical.
}
\examples{
  dgroups(num.groups = 2, probs = c(0.01, 0.2), num.series = 5)
  dgroups(num.groups = 2, probs = c(0.01, 0.2), num.series = 5, max.trials = 60)
  dgroups(num.groups = 2, probs = c(0.01, 0.2), num.series = 5, fail.criteria = 2)
  dgroups(num.groups = 5, probs = seq(0.01, 0.05, by = 0.01), num.series = 2) 

}
\seealso{
\code{\link{dseries}}
  \code{\link{dtrials}}
}
