% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dseries.R
\name{dseries}
\alias{dseries}
\title{Completes multiple series of simulated drop tests.}
\usage{
dseries(num.series, tag.group = FALSE, group = 0, ...)
}
\arguments{
\item{num.series}{Integer. Specifies how many series of drop tests to
simulate.}

\item{tag.group}{Logical. If \code{TRUE}, group number is included in
output data table. Default is \code{FALSE}.}

\item{group}{Integer. Assigns a group number to each simulated drop test
within the same series. Only if \code{tag.group} is \code{TRUE}. Default
is 0.}

\item{...}{Passes \code{q}, \code{max.trials}, \code{fail.criteria},
and \code{fail.criteria} to \code{dtrials}. All are optional except
\code{q}.}
}
\value{
Data table of multiple simulated drop tests. 
  Each row of the data.table represents one simulated drop test. 
  \strong{Note:} Only data table is supported at this time.
  
 Where:
 \strong{F_CRITERIA} is the failure criteria specified (default is 1).
 
 \strong{REACT} is the total number of simulated reactions (failures).
 
 \strong{NON_REACT} is the total number of simulated non-reactions
  (successes).
  
 \strong{TRIALS} is the number of simulated trials performed until failure 
  condition met. If the failure condition was not met then this value
  will always be equal to \strong{MAX_TRIALS}.
  
 \strong{MAX_TRIALS} is the maximum number of simulated trials to perform
  as specified (default is 20). \strong{TRIALS} will always be less than
  or equal to \strong{MAX_TRIALS}.
  
 \strong{PCT_REACT} is the percent of simulated trials that yielded a
  reaction (failure).
  
 \strong{Q} is the probability of failure (reaction) as specified.
 
 \strong{P} is the probability of success (non-reaction).
 
 \strong{RESULT} is whether the simulated test series as a whole failed
  or passed based on the failure criteria specified.
}
\description{
\code{dseries} returns a series of simulated drop tests. Where each test
consists of trials (drops).
}
\examples{
  dseries(num.series = 5, q = 0.05)
  dseries(num.series = 5, q = 0.05, max.trials = 60)
  dseries(num.series = 5, q = 0.05, max.trials = 60, fail.criteria = 2)
  
}
\seealso{
\code{\link{dtrials}}
}
