% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dseries.R
\name{dseries}
\alias{dseries}
\title{Completes multiple series of simulated drop tests.}
\usage{
dseries(num.series, tag.group = FALSE, group = 0, ...)
}
\arguments{
\item{num.series}{Integer. Specifies how many series of drop tests to
simulate.}

\item{tag.group}{Logical. If \code{TRUE}, group number is included in
output data table. Default is \code{FALSE}.}

\item{...}{Passes \code{q}, \code{max.trials}, \code{fail.criteria},
\code{raw.data}, and \code{fail.criteria} to \code{dtrials}. All are
optional except \code{q}.}

\item{group:}{Integer. Assigns a group number to each simulated drop test
within the same series. Only if \code{tag.group} is \code{TRUE}. Default
is 0.}
}
\value{
Data table of multiple simulated drop tests. 
  Where each row of the data.table represents one simulated drop test. 
  \strong{Note:} Only data table is supported at this time.
}
\description{
\code{dseries} returns a series of simulated drops tests. Where each test
consists of independent drops (trials).
}
