% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtrials.R
\name{dtrials}
\alias{dtrials}
\title{Completes a simulated drop test.}
\usage{
dtrials(q, max.trials = 20, fail.criteria = 1, raw.data = FALSE,
  data.structure = "data.table")
}
\arguments{
\item{q}{Integer. Specifies the probability that a reaction occurs. 
A reaction is interpreted as a failure therfore q is the probability
of failure for the bernoulli trials. Probability of success would be
\eqn{p = 1 - q}. See \url{https://en.wikipedia.org/wiki/Bernoulli_trial}}

\item{max.trials}{Integer. The maximum number of bernoulli trials to
perform. where each trial represents one drop of the impactor onto
a sample. The simulated test can only reach this number of trials if
no reactions (failures) occur. Default is 20.}

\item{fail.criteria}{Integer. Specifies number of reactions (failures)
that can occur before an entire test is considered a failure. Default
is 1.}

\item{data.structure}{Instructs function to return result as a data.table
or a list. Default is data.table. 

\strong{Note:} Other functions in this package only work with data.tables. List is
an option strictly for future flexibility.}
}
\value{
Result of simulation as a data table or list -- depending on value
  of \code{data.structure}. Where #@%#%#@^#@$^#@$^
}
\description{
\code{dtrials} returns bernoulli trials that make up a drop test.
}
\examples{
  dtrials(0.05)
  dtrials(0.05, max.trials = 60)
  dtrials(0.05, fail.criteria = 2)
  dtrials(0.05, max.trials = 60, fail.criteria = 2)
  dtrials(0.05, data.structure = "list")    
  
\dontrun{
dtrials()
}
}
