% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtrials.R
\name{dtrials}
\alias{dtrials}
\title{Completes a simulated drop test.}
\usage{
dtrials(q, max.trials = 20, fail.criteria = 1,
  data.structure = "data.table")
}
\arguments{
\item{q}{Integer. Specifies the probability that a reaction occurs. 
A reaction is interpreted as a failure therfore q is the probability
of failure for the bernoulli trials. Probability of success would be
\eqn{p = 1 - q}. See \url{https://en.wikipedia.org/wiki/Bernoulli_trial}}

\item{max.trials}{Integer. The maximum number of bernoulli trials to
perform. where each trial represents one drop of the impactor onto
a sample. Trials performed will always be less than or equal to
\code{max.trials}. Default is 20.}

\item{fail.criteria}{Integer. Specifies number of reactions (failures)
that can occur before an entire test is considered a failure. Must be 
less than or equal to \code{max.trials}. Default is 1.}

\item{data.structure}{Instructs function to return result as a data.table
or list. Default is data.table. 

\strong{Note:} Other functions in this package only work with
data.tables. List is an option strictly for future flexibility.}
}
\value{
A data table or list. Containing the following elements:
  
 \itemize{ 
  \item \strong{F_CRITERIA} The failure criteria specified.
   
  \item \strong{REACT} The total number of simulated reactions
   (failures).
   
  \item \strong{NON_REACT} The total number of simulated non-reactions
    (successes).
    
  \item \strong{TRIALS} The number of simulated trials performed until
    the failure condition was met.
         
   \item \strong{MAX_TRIALS} The maximum number of simulated trials
    specified.
    
   \item \strong{PCT_REACT} The percent of simulated trials that yielded
    a reaction (failure).
    
   \item \strong{Q} The probability of failure (reaction) as specified.
   
   \item \strong{P} The probability of success (non-reaction).
   
   \item \strong{RESULT} Whether the simulated test series as a whole
   failed or passed based on the failure criteria specified.
 }
}
\description{
\code{dtrials} returns simulated bernoulli trials (drops) that compose one
  drop test.
}
\examples{
  dtrials(0.05)
  dtrials(0.05, max.trials = 60)
  dtrials(0.05, fail.criteria = 2)
  dtrials(0.05, max.trials = 60, fail.criteria = 2)
  dtrials(0.05, data.structure = "list")
  
}
\seealso{
\code{\link{droptest}}
}
\author{
Chad Ross \email{chad.ross@gmail.com}
}
